Refactor line center
